Improve Auth Schema definition
There was a review regarding not trimming some of the fields so I did that to all fields
The email's field descirption shouldn't be "required" as it is expected to be required by default in the first place.
It is specifically "Invalid email address" as we are performing an additional check to verify the text format matching that of an email.