Create Post list page
As one of the issues requested that a post list page be created, I created this in that regard.
This shows the posts made by people in the explore route of the application. A post can have:
- Creator Name (first and last)
- A form of Profile picture if supported else we just default to initials from the full name
- I displayed the rating of the product, and the total number of reviews that has amounted to the rating
- I display a descripton about what the post is about
- I added a price tag for the product, and also specified if there was any discount going on
- If the content had some kind of duration that can be associated with it, there is also a case for including that in the post as well

This page also contains filter options, so that users can filter the posts they see based on their interests
- We also have a search bar for easy search of products.