Make Logging more Robust
I realised that using Winston alone as the source of truth for logging
was not going to work, as client-side javascript doesn't have access to
fs, os, path, and these kins of modules that are necessary for Winston to function well.

I encountered a couple of errors relating to the above, that made me consider a different route.

What i ended up with is:
1. Use Winston for only server side code
2. Use the regular logger for anything client side

I also tried to make the custom Logger more robust to convey more information.