Implement PostHog with Paystack
PostHog is a really cheap and robust tool to manage interaction and analytics
in a frotend application.

I configured it initially, mainly for managing analytics on user payment. i.e

Every time there is a successful or Failed payment so that, the developer can be aware
and necessary steps can be made to help in such situations

Also it just makes sense to add this not for this reason only, but for general analytics of the application, making the app
as "production-grade" as possible.