Handling EOL characters for various OS
There seems to be a know issue with how EOL (line breaks) characters are process on both Unix based
systems and Windows.

In Unix based system it is interpreted as LF (\n) Line Feed, but on Windows system
it is interpreted as CRLF (\r\n) two sequence: Carriage Return Line Feed.

And from research I have found out that It is easier for windows to handle LF sequence
but difficult for Unix based system to properly handle CRLF and so the solution I found to such a problem
is to add this .gitattribute with the content in it.

This ensure that Files that are created or edited on Windows machine do not end up chaning the EOL character from LF if that is what it was previously
or creating it's own line breaks as CRLF instead of LF.

This ulitmately unifies the use of LF as some kind of standard for new line characters.\

Finally, setting LF as the line endings for the whole repo would avoid strange things that happen when on Windows,
 Prettier retains its default of LF but Git and your code editor continue to use their default of CRLF for line endings: