Fix all instance of isFree contradition.
During the review from copilot, it pointed out contradictions with the isFree and price property of the content object.

This issues arose bacuse I changed all property name from isPremium to isFree, and didnot not also invert the values since
they meant two oppositite thinking of it in English terms.

In this commit I have fixed all of those issues.