Fix password trim issue
Copilot alerted me in the review that passphrases could have spaced at the end or something.

If the user is allowed to innput the password twice, then why trim it if they include space at the end twice?
I read some post on stack overflow, and there seems to be a large category of people that believe users should be able to choose whatever password they want, and also considering
the fact that they have to input it twice, and we also check the passwords to see whether they are the same.