Add utility function to generate Formdata from object
This was necessary because the wway that the upload form is structured, with respect to the multi-step format,
and only displaying certain step based on progress.

The way that we are implementing this is that we are removing the elements from the UI for each step, such that when the user gets to the end of the form, and wants to submit,
there will be an issue using react "server actions" beacuse It automatically check fors all input elements in the form
and gets that value and stores them in a form data based on their name attribute.

The way to work around this is to use the oject state that was used to ensure controlled inputs for the form data.

And since the "server action" function required a form data as an argument, I needed this function to do the conversion.