Define utilities required for getting content and comments by id
In this we also saw a usecase of reusing the transformcontent function for both the explore page and
the checkout page, since the structure of both response are similar.

I initially had it in mind to cache the reponse for the checkout page, but there might occur an issue where user's commend might not update
on refresh of the page appropriately due to heavy caching, so I thought it to be best to avoid that for now at least.