Set up unit tests.
This is verfy essential as previoulsy I have found myself dealing the pain of
managing arithmetic in javascript and hoping that it works. In this case I can
be sure to test for edge cases and ensure that everything works as expected.

Also testing the debouce to see that it works together with the form validation
was satisfying