Attending to requested changes on test file
Previously I attempted to load the variable for the email and password from the env variables
but I didn't want to install additional libraries just for the sake of loading envs alone.

But I found that implicitly Nextjs uses the next/env package to manage loading envs into some parts of its environment, so I just opted for that package.
It seems lightweight and has done its job perfectly.