Remove dummy data & eidt author type
Previously I was receiving the author types as a string(this was supposed to mean the author id).
But that was not so useful to mea as I wanted the author data: firstname, lastname, and the likes.

This new changed reuired that I edited some of my previous way of consuing that field.