Improve Comment management.
Previouls we had dummy states for the commenter name and the time commented.

Part of this was because I was not receiving these details from the backend. Since that has been resolved now, I needed some refactor to get things
working as expected.

I also disabled the links to pages that we do not have available yet. This is just to prevent being redirected to a 404 page.
Keeping things professional here...